Fixed goods

+++ Atlas.Presentation/Atlas.WebApi/Models/UpdateGoodDto.cs
@@ -25,7 +25,7 @@ namespace Atlas.WebApi.Models
 
         public float Volume { get; set; }
 
-        public int Discount { get; set; }
+        public float Discount { get; set; }
 
         public void Mapping(Profile profile)
         {
